refactor: move generator
